requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
Faker==19.6.2
fastapi==0.103.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.1
pluggy==1.3.0
postgres==4.0
psycopg2-binary==2.9.7
psycopg2-pool==1.1
pydantic==2.3.0
pydantic_core==2.6.3
pytest==7.4.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.5
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
